+++ R/cc_brightness.R
@@ -183,7 +183,7 @@ cc_brightness_timeseries <- function(img, frames_per_set,
   if (length(detrend) == 1) detrend %<>% rep(2)
   thresh %<>% prepare_thresh()
   filt %<>% prepare_filt()
-  if (is.character(img)) img %<>% ijtiff::read_tif()
+  if (is.character(img)) img %<>% ijtiff::read_tif(msg = FALSE)
   checkmate::assert_array(img, d = 4)
   if (dim(img)[4] < frames_per_set) {
     custom_stop("
@@ -313,7 +313,7 @@ cc_brightness_file <- function(path, ch1 = 1, ch2 = 2,
   path %<>% filesstrings::before_last_dot() %>%
     paste0("cc_brightness", "/", ., make_cc_nb_filename_ending(cc_b)) %>%
     deduplicate_cc_nb_filename()
-  ijtiff::write_tif(cc_b, path)
+  ijtiff::write_tif(cc_b, path, msg = FALSE)
 }
 
 cc_brightness_timeseries_file <- function(path, frames_per_set,
@@ -346,7 +346,7 @@ cc_brightness_timeseries_file <- function(path, frames_per_set,
       make_cc_nb_filename_ending(cc_b_ts)
     ) %>%
     deduplicate_cc_nb_filename()
-  ijtiff::write_tif(cc_b_ts, path)
+  ijtiff::write_tif(cc_b_ts, path, msg = FALSE)
 }
 
 #' Cross-correlated brightness calculations for every image in a folder.